src/02_mcp/p1_mcp_agent/agent.py
import datetime
from zoneinfo import ZoneInfo
from google.adk.agents import Agent

import arxiv
import json
import os
from typing import List
from google.adk.models.lite_llm import LiteLlm
from google.adk.tools.mcp_tool.mcp_toolset import MCPToolset
from google.adk.tools.mcp_tool.mcp_session_manager import StdioConnectionParams, StreamableHTTPConnectionParams
from google.adk.tools.mcp_tool.mcp_session_manager import SseServerParams
from mcp import StdioServerParameters
from dotenv import load_dotenv

def search_papers(topic: str, max_results: int = 5) -> List[str]:
    """
    Search for papers on arXiv based on a topic and store their information.
    
    Args:
        topic: The topic to search for
        max_results: Maximum number of results to retrieve (default: 5)
        
    Returns:
        List of paper IDs found in the search
    """
    
    # Use arxiv to find the papers 
    client = arxiv.Client()

    # Search for the most relevant articles matching the queried topic
    search = arxiv.Search(
        query = topic,
        max_results = max_results,
        sort_by = arxiv.SortCriterion.Relevance
    )

    papers = client.results(search)
    
    # Create directory for this topic
    path = os.path.join(PAPER_DIR, topic.lower().replace(" ", "_"))
    os.makedirs(path, exist_ok=True)
    
    file_path = os.path.join(path, "papers_info.json")

    # Try to load existing papers info
    try:
        with open(file_path, "r") as json_file:
            papers_info = json.load(json_file)
    except (FileNotFoundError, json.JSONDecodeError):
        papers_info = {}

    # Process each paper and add to papers_info  
    paper_ids = []
    for paper in papers:
        paper_ids.append(paper.get_short_id())
        paper_info = {
            'title': paper.title,
            'authors': [author.name for author in paper.authors],
            'summary': paper.summary,
            'pdf_url': paper.pdf_url,
            'published': str(paper.published.date())
        }
        papers_info[paper.get_short_id()] = paper_info
    
    # Save updated papers_info to json file
    with open(file_path, "w") as json_file:
        json.dump(papers_info, json_file, indent=2)
    
    print(f"Results are saved in: {file_path}")
    
    return paper_ids

def extract_info(paper_id: str) -> str:
    """
    Search for information about a specific paper across all topic directories.
    
    Args:
        paper_id: The ID of the paper to look for
        
    Returns:
        JSON string with paper information if found, error message if not found
    """
 
    for item in os.listdir(PAPER_DIR):
        item_path = os.path.join(PAPER_DIR, item)
        if os.path.isdir(item_path):
            file_path = os.path.join(item_path, "papers_info.json")
            if os.path.isfile(file_path):
                try:
                    with open(file_path, "r") as json_file:
                        papers_info = json.load(json_file)
                        if paper_id in papers_info:
                            return json.dumps(papers_info[paper_id], indent=2)
                except (FileNotFoundError, json.JSONDecodeError) as e:
                    print(f"Error reading {file_path}: {str(e)}")
                    continue
    
    return f"There's no saved information related to paper {paper_id}."

use_model = "deepseek"

if use_model == "deepseek":
    model = LiteLlm(model="deepseek/deepseek-chat")
if use_model == "gpt-4o":
    model = LiteLlm(model="azure/gpt-4o")


# 加载 .env 文件中的环境变量
load_dotenv()
# https://www.tavily.com/free-api-key
# https://mcp.tavily.com/mcp/?tavilyApiKey=<your-api-key>
tavily_api_key = os.getenv("TAVILY_API_KEY")
tavily_api_url = f"https://mcp.tavily.com/mcp/?tavilyApiKey={tavily_api_key}"
tavily_mcp = MCPToolset(
    connection_params=StreamableHTTPConnectionParams(
        url=tavily_api_url,
        timeout=5,
    ),
)
# tavily_mcp = MCPToolset(
#     connection_params=StdioConnectionParams(
#         server_params=StdioServerParameters(
#             command='npx',
#             args=[  "-y","mcp-remote", tavily_api_url],
#         ),
#         timeout=50,
#     ),
# )

# https://github.com/34892002/bilibili-mcp-js
bilibili_mcp = MCPToolset(
    connection_params=StdioConnectionParams(
        server_params=StdioServerParameters(
            command='npx',
            args=['bilibili-mcp'],
        ),
    ),
)



root_agent = Agent(
    name="search_agent",
    model=model,
    description=(
        "Agent to answer questions about what you want to search."
    ),
    instruction=(
        "You are a helpful agent who can answer user questions about what you want to search."
    ),
    # tools=[get_weather, get_current_time, search_papers, extract_info],
    tools=[tavily_mcp, bilibili_mcp],
)