pyproject.toml
[project]
name = "adk-get-started"
version = "0.1.0"
description = ""
authors = [
    {name = "wyzBelinda",email = "wyzBelinda@163.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.12.0",
    "google-adk>=1.12.0",
    "jupyter>=1.1.1",
    "litellm>=1.76.0",
]

[tool.uv.workspace]
members = [
    "origene_langraph",
    "tests/adk-mcp-a2a-crash-course",
]

[dependency-groups]
dev = [
    "arxiv>=2.2.0",
]
